+++ cmake/ExternLibs.cmake
@@ -32,6 +32,11 @@ function(ADDQT5NETWORK TARGET_NAME)
 	target_link_libraries(${TARGET_NAME} PRIVATE Qt5::Network)
 endfunction()
 
+function(ADDSPDLOG TARGET_NAME)
+	find_package(spdlog CONFIG REQUIRED)
+    target_link_libraries(${TARGET_NAME} PRIVATE spdlog::spdlog)
+endfunction()
+
 function(FETCHSHAREDTYPES)
 	FetchContent_Declare(
 	  shared_types